Move some helper functions into `util::test_app`
