+++ src/tests/all.rs
@@ -21,18 +21,13 @@ use cargo_registry::{
         EncodableCategory, EncodableCategoryWithSubcategories, EncodableCrate, EncodableKeyword,
         EncodableOwner, EncodableVersion, GoodCrate,
     },
-    App, Config, Env, Replica, Uploader,
 };
 use std::{
     borrow::Cow,
-    sync::{
-        atomic::{AtomicUsize, Ordering},
-        Arc,
-    },
+    sync::atomic::{AtomicUsize, Ordering},
 };
 
 use diesel::prelude::*;
-use reqwest::{blocking::Client, Proxy};
 
 mod account_lock;
 mod authentication;
@@ -101,62 +96,6 @@ pub struct CategoryMeta {
 #[derive(Deserialize)]
 pub struct OkBool {
     ok: bool,
-}
-
-fn simple_config() -> Config {
-    let uploader = Uploader::S3 {
-        bucket: s3::Bucket::new(
-            String::from("alexcrichton-test"),
-            None,
-            dotenv::var("S3_ACCESS_KEY").unwrap_or_default(),
-            dotenv::var("S3_SECRET_KEY").unwrap_or_default(),
-            // When testing we route all API traffic over HTTP so we can
-            // sniff/record it, but everywhere else we use https
-            "http",
-        ),
-        cdn: None,
-    };
-
-    Config {
-        uploader,
-        session_key: "test this has to be over 32 bytes long".to_string(),
-        gh_client_id: dotenv::var("GH_CLIENT_ID").unwrap_or_default(),
-        gh_client_secret: dotenv::var("GH_CLIENT_SECRET").unwrap_or_default(),
-        gh_base_url: "http://api.github.com".to_string(),
-        db_url: env("TEST_DATABASE_URL"),
-        replica_db_url: None,
-        env: Env::Test,
-        max_upload_size: 3000,
-        max_unpack_size: 2000,
-        mirror: Replica::Primary,
-        // When testing we route all API traffic over HTTP so we can
-        // sniff/record it, but everywhere else we use https
-        api_protocol: String::from("http"),
-        publish_rate_limit: Default::default(),
-        blocked_traffic: Default::default(),
-        domain_name: "crates.io".into(),
-        allowed_origins: Vec::new(),
-    }
-}
-
-fn build_app(
-    config: Config,
-    proxy: Option<String>,
-) -> (Arc<App>, conduit_middleware::MiddlewareBuilder) {
-    let client = if let Some(proxy) = proxy {
-        let mut builder = Client::builder();
-        builder = builder
-            .proxy(Proxy::all(&proxy).expect("Unable to configure proxy with the provided URL"));
-        Some(builder.build().expect("TLS backend cannot be initialized"))
-    } else {
-        None
-    };
-
-    let app = App::new(config, client);
-    assert_ok!(assert_ok!(app.primary_database.get()).begin_test_transaction());
-    let app = Arc::new(app);
-    let handler = cargo_registry::build_handler(Arc::clone(&app));
-    (app, handler)
 }
 
 // Return the environment variable only if it has been defined
